lint error and diff cleanup
