+++ spotlight-client/package.json
@@ -19,6 +19,7 @@
     "@reach/router": "^1.3.4",
     "@types/classnames": "^2.2.11",
     "@types/d3-array": "^2.8.0",
+    "@types/d3-color": "^2.0.1",
     "@types/d3-format": "^2.0.0",
     "@types/d3-interpolate": "^2.0.0",
     "@types/d3-scale": "^3.2.2",
@@ -35,6 +36,7 @@
     "change-case": "^4.1.2",
     "classnames": "^2.2.6",
     "d3-array": "^2.9.1",
+    "d3-color": "^2.0.0",
     "d3-format": "^2.0.0",
     "d3-interpolate": "^2.0.1",
     "d3-scale": "^3.2.3",
