+++ spotlight-client/src/charts/ColorLegend.tsx
@@ -18,8 +18,9 @@
 import classNames from "classnames";
 import React from "react";
 import styled from "styled-components/macro";
-import { animation, highlightFade } from "../UiLibrary";
+import { animation } from "../UiLibrary";
 import { ItemToDisplay, ItemToHighlight } from "./types";
+import { highlightFade } from "./utils";
 
 const ColorLegendWrapper = styled.div`
   display: flex;
