+++ spotlight-client/src/charts/WindowedTimeSeries.tsx
@@ -21,10 +21,10 @@ import { format, isEqual } from "date-fns";
 import React, { useCallback, useEffect, useState } from "react";
 import MinimapXYFrame from "semiotic/lib/MinimapXYFrame";
 import styled from "styled-components/macro";
-import { animation, colors, highlightFade } from "../UiLibrary";
+import { animation, colors } from "../UiLibrary";
 import { formatAsNumber } from "../utils";
 import BaseChartWrapper from "./ChartWrapper";
-import { getDataWithPct } from "./utils";
+import { getDataWithPct, highlightFade } from "./utils";
 import ColorLegend from "./ColorLegend";
 import XHoverController from "./XHoverController";
 import { HistoricalPopulationBreakdownRecord } from "../metricsApi";
